requirements.txt
torch
pandas
numpy
torchvision
Pillow
matplotlib
optree
sqlite3